refactor: rearrange calling classes
