+++ app/src/main/kotlin/com/wire/android/ui/calling/ongoing/participantsview/horizentalview/CallingHorizentalView.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.common
+package com.wire.android.ui.calling.ongoing.participantsview.horizentalview
 
 import android.view.View
 import androidx.compose.animation.core.tween
@@ -16,9 +16,9 @@ import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.wire.android.ui.calling.ConversationName
-import com.wire.android.ui.calling.ParticipantTile
 import com.wire.android.ui.calling.getConversationName
 import com.wire.android.ui.calling.model.UICallParticipant
+import com.wire.android.ui.calling.ongoing.participantsview.ParticipantTile
 import com.wire.android.ui.common.dimensions
 import com.wire.android.ui.theme.wireDimensions
