+++ app/src/test/kotlin/com/wire/android/ui/calling/OngoingCallViewModelTest.kt
@@ -2,6 +2,7 @@ package com.wire.android.ui.calling
 
 import androidx.lifecycle.SavedStateHandle
 import com.wire.android.navigation.NavigationManager
+import com.wire.android.ui.calling.ongoing.OngoingCallViewModel
 import com.wire.kalium.logic.data.id.QualifiedIdMapper
 import com.wire.kalium.logic.feature.call.usecase.ObserveEstablishedCallsUseCase
 import com.wire.kalium.logic.feature.call.usecase.RequestVideoStreamsUseCase
